+++ src/main/java/edu/swarthmore/cs/cs71/swatify/Application.java
@@ -1,6 +1,9 @@
 package edu.swarthmore.cs.cs71.swatify;
 
+import edu.swarthmore.cs.cs71.swatify.controllers.AlbumsController;
+import edu.swarthmore.cs.cs71.swatify.controllers.ArtistsController;
 import edu.swarthmore.cs.cs71.swatify.controllers.UsersController;
+import edu.swarthmore.cs.cs71.swatify.models.Artist;
 
 import static spark.Spark.*;
 
@@ -9,14 +12,11 @@ public class Application {
         setRoutes();
     }
 
-    private static void setRoutes() {
+    public static void setRoutes() {
         path("/api/v1", () -> {
-            path("/users", () -> {
-                get("/*", UsersController::getUser);
-                put("/*", UsersController::createUser);
-                patch("/*", UsersController::updateUser);
-                delete("/*", UsersController::deleteUser);
-            });
+            UsersController usersController = new UsersController();
+            ArtistsController artistsController = new ArtistsController();
+            AlbumsController albumsController = new AlbumsController();
         });
     }
 }