+++ src/main/java/edu/swarthmore/cs/cs71/swatify/controllers/UsersController.java
@@ -1,27 +1,46 @@
 package edu.swarthmore.cs.cs71.swatify.controllers;
 
 import edu.swarthmore.cs.cs71.swatify.models.User;
+import edu.swarthmore.cs.cs71.swatify.util.GsonUtil;
 import edu.swarthmore.cs.cs71.swatify.util.HibernateUtil;
 import spark.Request;
 import spark.Response;
 
+import static spark.Spark.*;
+
 public class UsersController {
-    public static User getUser(Request request, Response response) {
-        int userId = Integer.parseInt(request.params("id"));
-        return HibernateUtil.getObjectById(User.class, userId);
+    public UsersController() {
+        path("/users", () -> {
+            get("/:id", (request, response) -> getUser(Integer.parseInt(request.params("id"))), GsonUtil::toJson);
+
+            post("", (request, response) -> createUser(GsonUtil.fromJson(User.class, request.body())),  GsonUtil::toJson);
+
+            patch("/:id", (request, response) -> updateUser(GsonUtil.fromJson(User.class, request.body())), GsonUtil::toJson);
+
+            delete("/:id", (request, response) -> deleteUser(Integer.parseInt(request.params("id"))), GsonUtil::toJson);
+
+            after((req, res) -> res.type("application/json"));
+
+            exception(IllegalArgumentException.class, (e, req, res) -> {
+                res.status(400);
+            });
+        });
     }
 
-    public static User createUser(Request request, Response response) {
-        return new User("","");
+    public static User getUser(int id) {
+        return HibernateUtil.getObjectById(User.class, id);
     }
 
-    public static User updateUser(Request request, Response response) {
-        int id = Integer.parseInt(request.params("id"));
-        return new User("","");
+    public static User createUser(User user) {
+        return HibernateUtil.saveObject(user);
     }
 
-    public static User deleteUser(Request request, Response response) {
-        int id = Integer.parseInt(request.params("id"));
-        return new User("","");
+    public static User updateUser(User user) {
+        return HibernateUtil.updateObject(user);
     }
+
+    public static User deleteUser(int id) {
+        return HibernateUtil.deleteObject(User.class, id);
+    }
+
 }