+++ src/main/java/edu/swarthmore/cs/cs71/swatify/models/Post.java
@@ -1,0 +1,30 @@
+package edu.swarthmore.cs.cs71.swatify.models;
+
+import javax.persistence.*;
+
+@Entity
+@Table
+public class Post extends PublicAction {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private String content;
+
+    public Post() { }
+
+    public Post(String content, int userId) {
+        this.userId = userId;
+        this.content = content;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public void setContent(String content) { this.content = content; }
+}